@@ -16,23 +16,25 @@
         }
     ],
     "require": {
-        "php": "^5.3.3|^7.0",
+        "php": "^7.1",
         "friendsofsymfony/oauth2-php": "~1.1",
-        "symfony/framework-bundle": "~2.2|~3.0",
-        "symfony/security-bundle": "~2.1|~3.0",
-        "symfony/dependency-injection": "^2.0.5|~3.0"
+        "symfony/framework-bundle": "~3.0|^4.0",
+        "symfony/security-bundle": "~3.0|^4.0",
+        "symfony/dependency-injection": "^2.8|~3.0|^4.0"
     },
     "require-dev": {
-        "symfony/class-loader": "~2.1|~3.0",
-        "symfony/yaml": "~2.1|~3.0",
-        "symfony/form": "~2.3|~3.0",
-        "symfony/console": "~2.1|~3.0",
+        "symfony/class-loader": "~3.0|^4.0",
+        "symfony/yaml": "~3.0|^4.0",
+        "symfony/form": "~3.0|^4.0",
+        "symfony/console": "~3.0|^4.0",
+        "symfony/templating": "~3.0|^4.0",
         "willdurand/propel-typehintable-behavior": "^1.0.4",
         "propel/propel1": "^1.6.5",
         "phing/phing": "~2.4",
         "doctrine/mongodb-odm": "~1.0",
         "doctrine/doctrine-bundle": "~1.0",
-        "doctrine/orm": "~2.2"
+        "doctrine/orm": "~2.2",
+        "phpunit/phpunit": "~4.8|~5.0"
     },
     "suggest": {
         "doctrine/doctrine-bundle": "*",
@@ -48,7 +50,7 @@
     },
     "extra": {
         "branch-alias": {
-            "dev-master": "1.5-dev"
+            "dev-master": "2.0-dev"
         }
     }
 }
