@@ -1,5 +1,7 @@
 <?php
 
+declare(strict_types=1);
+
 /*
  * This file is part of the FOSOAuthServerBundle package.
  *
@@ -28,9 +30,11 @@ class Configuration implements ConfigurationInterface
     public function getConfigTreeBuilder()
     {
         $treeBuilder = new TreeBuilder();
+
+        /** @var ArrayNodeDefinition $rootNode */
         $rootNode = $treeBuilder->root('fos_oauth_server');
 
-        $supportedDrivers = array('orm', 'mongodb', 'propel');
+        $supportedDrivers = ['orm', 'mongodb', 'propel'];
 
         $rootNode
             ->children()
@@ -47,7 +51,8 @@ class Configuration implements ConfigurationInterface
                 ->scalarNode('refresh_token_class')->isRequired()->cannotBeEmpty()->end()
                 ->scalarNode('auth_code_class')->isRequired()->cannotBeEmpty()->end()
                 ->scalarNode('model_manager_name')->defaultNull()->end()
-            ->end();
+            ->end()
+        ;
 
         $this->addAuthorizeSection($rootNode);
         $this->addServiceSection($rootNode);
@@ -72,13 +77,14 @@ class Configuration implements ConfigurationInterface
                                 ->scalarNode('name')->defaultValue('fos_oauth_server_authorize_form')->cannotBeEmpty()->end()
                                 ->arrayNode('validation_groups')
                                     ->prototype('scalar')->end()
-                                    ->defaultValue(array('Authorize', 'Default'))
+                                    ->defaultValue(['Authorize', 'Default'])
                                 ->end()
                             ->end()
                         ->end()
                     ->end()
                 ->end()
-            ->end();
+            ->end()
+        ;
     }
 
     private function addServiceSection(ArrayNodeDefinition $node)
@@ -97,13 +103,14 @@ class Configuration implements ConfigurationInterface
                             ->scalarNode('auth_code_manager')->defaultValue('fos_oauth_server.auth_code_manager.default')->end()
                             ->arrayNode('options')
                                 ->useAttributeAsKey('key')
-                                ->treatNullLike(array())
+                                ->treatNullLike([])
                                 ->prototype('scalar')->end()
                             ->end()
                         ->end()
                     ->end()
                 ->end()
-            ->end();
+            ->end()
+        ;
     }
 
     private function addTemplateSection(ArrayNodeDefinition $node)
@@ -116,6 +123,7 @@ class Configuration implements ConfigurationInterface
                         ->scalarNode('engine')->defaultValue('twig')->end()
                     ->end()
                 ->end()
-            ->end();
+            ->end()
+        ;
     }
 }
