@@ -1,5 +1,7 @@
 <?php
 
+declare(strict_types=1);
+
 /*
  * This file is part of the FOSOAuthServerBundle package.
  *
@@ -50,7 +52,7 @@ class Authorize
      * @param bool  $accepted
      * @param array $query
      */
-    public function __construct($accepted, array $query = array())
+    public function __construct($accepted, array $query = [])
     {
         foreach ($query as $key => $value) {
             $this->{$key} = $value;
