@@ -1,5 +1,7 @@
 <?php
 
+declare(strict_types=1);
+
 /*
  * This file is part of the FOSOAuthServerBundle package.
  *
@@ -12,25 +14,24 @@
 namespace FOS\OAuthServerBundle\Storage;
 
 use FOS\OAuthServerBundle\Model\AccessTokenManagerInterface;
-use FOS\OAuthServerBundle\Model\RefreshTokenManagerInterface;
 use FOS\OAuthServerBundle\Model\AuthCodeManagerInterface;
-use FOS\OAuthServerBundle\Model\ClientManagerInterface;
 use FOS\OAuthServerBundle\Model\ClientInterface;
-use Symfony\Component\Security\Core\User\UserProviderInterface;
-use Symfony\Component\Security\Core\Encoder\EncoderFactoryInterface;
-use Symfony\Component\Security\Core\Exception\AuthenticationException;
+use FOS\OAuthServerBundle\Model\ClientManagerInterface;
+use FOS\OAuthServerBundle\Model\RefreshTokenManagerInterface;
+use OAuth2\IOAuth2GrantClient;
+use OAuth2\IOAuth2GrantCode;
+use OAuth2\IOAuth2GrantExtension;
+use OAuth2\IOAuth2GrantImplicit;
+use OAuth2\IOAuth2GrantUser;
+use OAuth2\IOAuth2RefreshTokens;
+use OAuth2\Model\IOAuth2Client;
 use OAuth2\OAuth2;
 use OAuth2\OAuth2ServerException;
-use OAuth2\IOAuth2RefreshTokens;
-use OAuth2\IOAuth2GrantUser;
-use OAuth2\IOAuth2GrantCode;
-use OAuth2\IOAuth2GrantImplicit;
-use OAuth2\IOAuth2GrantClient;
-use OAuth2\IOAuth2GrantExtension;
-use OAuth2\Model\IOAuth2Client;
+use Symfony\Component\Security\Core\Encoder\EncoderFactoryInterface;
+use Symfony\Component\Security\Core\Exception\AuthenticationException;
+use Symfony\Component\Security\Core\User\UserProviderInterface;
 
-class OAuthStorage implements IOAuth2RefreshTokens, IOAuth2GrantUser, IOAuth2GrantCode, IOAuth2GrantImplicit,
-    IOAuth2GrantClient, IOAuth2GrantExtension, GrantExtensionDispatcherInterface
+class OAuthStorage implements IOAuth2RefreshTokens, IOAuth2GrantUser, IOAuth2GrantCode, IOAuth2GrantImplicit, IOAuth2GrantClient, IOAuth2GrantExtension, GrantExtensionDispatcherInterface
 {
     /**
      * @var ClientManagerInterface
@@ -86,7 +87,7 @@ class OAuthStorage implements IOAuth2RefreshTokens, IOAuth2GrantUser, IOAuth2Gra
         $this->userProvider = $userProvider;
         $this->encoderFactory = $encoderFactory;
 
-        $this->grantExtensions = array();
+        $this->grantExtensions = [];
     }
 
     /**
@@ -167,9 +168,9 @@ class OAuthStorage implements IOAuth2RefreshTokens, IOAuth2GrantUser, IOAuth2Gra
             $encoder = $this->encoderFactory->getEncoder($user);
 
             if ($encoder->isPasswordValid($user->getPassword(), $password, $user->getSalt())) {
-                return array(
+                return [
                     'data' => $user,
-                );
+                ];
             }
         }
 
