@@ -1,5 +1,7 @@
 <?php
 
+declare(strict_types=1);
+
 /*
  * This file is part of the FOSOAuthServerBundle package.
  *
@@ -11,10 +13,10 @@
 
 namespace FOS\OAuthServerBundle\Tests\Document;
 
-use FOS\OAuthServerBundle\Document\TokenManager;
 use FOS\OAuthServerBundle\Document\AccessToken;
+use FOS\OAuthServerBundle\Document\TokenManager;
 
-class TokenManagerTest extends \PHPUnit_Framework_TestCase
+class TokenManagerTest extends \PHPUnit\Framework\TestCase
 {
     protected $class;
     protected $dm;
@@ -28,26 +30,36 @@ class TokenManagerTest extends \PHPUnit_Framework_TestCase
         }
 
         $this->class = 'FOS\OAuthServerBundle\Document\AccessToken';
-        $this->repository = $this->getMock('Doctrine\ODM\MongoDB\DocumentRepository', array(), array(), '', false);
-        $this->dm = $this->getMock('Doctrine\ODM\MongoDB\DocumentManager', array(), array(), '', false);
+        $this->repository = $this->getMockBuilder('Doctrine\ODM\MongoDB\DocumentRepository')
+            ->disableOriginalConstructor()
+            ->getMock()
+        ;
+        $this->dm = $this->getMockBuilder('Doctrine\ODM\MongoDB\DocumentManager')
+            ->disableOriginalConstructor()
+            ->getMock()
+        ;
         $this->dm->expects($this->once())
             ->method('getRepository')
             ->with($this->class)
-            ->will($this->returnValue($this->repository));
+            ->will($this->returnValue($this->repository))
+        ;
 
         $this->manager = new TokenManager($this->dm, $this->class);
     }
 
     public function testFindTokenByToken()
     {
+        /** @var \PHPUnit_Framework_MockObject_MockObject|TokenManager $manager */
         $manager = $this->getMockBuilder('FOS\OAuthServerBundle\Document\TokenManager')
             ->disableOriginalConstructor()
-            ->setMethods(array('findTokenBy'))
-            ->getMock();
+            ->setMethods(['findTokenBy'])
+            ->getMock()
+        ;
 
         $manager->expects($this->once())
             ->method('findTokenBy')
-            ->with($this->equalTo(array('token' => '1234')));
+            ->with($this->equalTo(['token' => '1234']))
+        ;
 
         $manager->findTokenByToken('1234');
     }
@@ -58,10 +70,12 @@ class TokenManagerTest extends \PHPUnit_Framework_TestCase
 
         $this->dm->expects($this->once())
             ->method('persist')
-            ->with($token);
+            ->with($token)
+        ;
         $this->dm->expects($this->once())
             ->method('flush')
-            ->with();
+            ->with()
+        ;
 
         $this->manager->updateToken($token);
     }
