@@ -1,5 +1,7 @@
 <?php
 
+declare(strict_types=1);
+
 /*
  * This file is part of the FOSOAuthServerBundle package.
  *
@@ -11,12 +13,12 @@
 
 namespace FOS\OAuthServerBundle\Tests\Form\Type;
 
-use FOS\OAuthServerBundle\Form\Type\AuthorizeFormType;
 use FOS\OAuthServerBundle\Form\Model\Authorize;
+use FOS\OAuthServerBundle\Form\Type\AuthorizeFormType;
 use FOS\OAuthServerBundle\Util\LegacyFormHelper;
-use Symfony\Component\Form\Test\TypeTestCase;
 use Symfony\Component\Form\FormBuilder;
 use Symfony\Component\Form\Forms;
+use Symfony\Component\Form\Test\TypeTestCase;
 
 class AuthorizeFormTypeTest extends TypeTestCase
 {
@@ -26,7 +28,8 @@ class AuthorizeFormTypeTest extends TypeTestCase
 
         $this->factory = Forms::createFormFactoryBuilder()
         ->addTypes($this->getTypes())
-        ->getFormFactory();
+        ->getFormFactory()
+        ;
 
         $this->builder = new FormBuilder(null, null, $this->dispatcher, $this->factory);
     }
@@ -34,13 +37,13 @@ class AuthorizeFormTypeTest extends TypeTestCase
     public function testSubmit()
     {
         $accepted = 'true';
-        $formData = array(
-            'client_id'     => '1',
+        $formData = [
+            'client_id' => '1',
             'response_type' => 'code',
-            'redirect_uri'  => 'http:\\localhost\test.php',
-            'state'         => 'testState',
-            'scope'         => 'testScope',
-        );
+            'redirect_uri' => 'http:\\localhost\test.php',
+            'state' => 'testState',
+            'scope' => 'testScope',
+        ];
 
         $authorize = new Authorize($accepted, $formData);
 
@@ -49,8 +52,8 @@ class AuthorizeFormTypeTest extends TypeTestCase
         $form->submit($formData);
 
         $this->assertTrue($form->isSynchronized());
-        $this->assertEquals($authorize, $form->getData());
-        $this->assertEquals((bool) $accepted, $authorize->accepted);
+        $this->assertSame($authorize, $form->getData());
+        $this->assertSame((bool) $accepted, $authorize->accepted);
 
         $view = $form->createView();
         $children = $view->children;
@@ -62,8 +65,8 @@ class AuthorizeFormTypeTest extends TypeTestCase
 
     protected function getTypes()
     {
-        return  array(
+        return  [
             new AuthorizeFormType(),
-        );
+        ];
     }
 }
