@@ -1,5 +1,7 @@
 <?php
 
+declare(strict_types=1);
+
 /*
  * This file is part of the FOSOAuthServerBundle package.
  *
@@ -11,11 +13,11 @@
 
 namespace FOS\OAuthServerBundle\Tests\Security\Authentication\Provider;
 
+use FOS\OAuthServerBundle\Model\AccessToken;
 use FOS\OAuthServerBundle\Security\Authentication\Provider\OAuthProvider;
 use FOS\OAuthServerBundle\Security\Authentication\Token\OAuthToken;
-use FOS\OAuthServerBundle\Model\AccessToken;
 
-class OAuthProviderTest extends \PHPUnit_Framework_TestCase
+class OAuthProviderTest extends \PHPUnit\Framework\TestCase
 {
     protected $user;
     protected $userProvider;
@@ -25,10 +27,23 @@ class OAuthProviderTest extends \PHPUnit_Framework_TestCase
 
     public function setUp()
     {
-        $this->user = $this->getMock('Symfony\Component\Security\Core\User\UserInterface');
-        $this->userProvider = $this->getMock('Symfony\Component\Security\Core\User\UserProviderInterface');
-        $this->serverService = $this->getMock('OAuth2\OAuth2', array('verifyAccessToken'), array(), '', false);
-        $this->userChecker = $this->getMock('Symfony\Component\Security\Core\User\UserCheckerInterface');
+        $this->user = $this->getMockBuilder('Symfony\Component\Security\Core\User\UserInterface')
+            ->disableOriginalConstructor()
+            ->getMock()
+        ;
+        $this->userProvider = $this->getMockBuilder('Symfony\Component\Security\Core\User\UserProviderInterface')
+            ->disableOriginalConstructor()
+            ->getMock()
+        ;
+        $this->serverService = $this->getMockBuilder('OAuth2\OAuth2')
+            ->disableOriginalConstructor()
+            ->setMethods(['verifyAccessToken'])
+            ->getMock()
+        ;
+        $this->userChecker = $this->getMockBuilder('Symfony\Component\Security\Core\User\UserCheckerInterface')
+            ->disableOriginalConstructor()
+            ->getMock()
+        ;
         $this->provider = new OAuthProvider($this->userProvider, $this->serverService, $this->userChecker);
     }
 
@@ -39,7 +54,8 @@ class OAuthProviderTest extends \PHPUnit_Framework_TestCase
 
         $this->user->expects($this->once())
             ->method('getRoles')
-            ->will($this->returnValue(array('ROLE_USER')));
+            ->will($this->returnValue(['ROLE_USER']))
+        ;
 
         $accessToken = new AccessToken();
         $accessToken->setUser($this->user);
@@ -47,17 +63,18 @@ class OAuthProviderTest extends \PHPUnit_Framework_TestCase
         $this->serverService->expects($this->once())
             ->method('verifyAccessToken')
             ->with('x')
-            ->will($this->returnValue($accessToken));
+            ->will($this->returnValue($accessToken))
+        ;
 
         $result = $this->provider->authenticate($token);
 
         $this->assertSame($this->user, $result->getUser());
-        $this->assertEquals($token->getToken(), $result->getToken());
+        $this->assertSame($token->getToken(), $result->getToken());
         $this->assertTrue($result->isAuthenticated());
         $this->assertCount(1, $result->getRoles());
 
         $roles = $result->getRoles();
-        $this->assertEquals('ROLE_USER', $roles[0]->getRole());
+        $this->assertSame('ROLE_USER', $roles[0]->getRole());
     }
 
     public function testAuthenticateReturnsTokenIfValidEvenIfNullData()
@@ -70,7 +87,8 @@ class OAuthProviderTest extends \PHPUnit_Framework_TestCase
         $this->serverService->expects($this->once())
             ->method('verifyAccessToken')
             ->with('x')
-            ->will($this->returnValue($accessToken));
+            ->will($this->returnValue($accessToken))
+        ;
 
         $result = $this->provider->authenticate($token);
 
@@ -90,7 +108,8 @@ class OAuthProviderTest extends \PHPUnit_Framework_TestCase
         $this->serverService->expects($this->once())
             ->method('verifyAccessToken')
             ->with('x')
-            ->will($this->returnValue($accessToken));
+            ->will($this->returnValue($accessToken))
+        ;
 
         $result = $this->provider->authenticate($token);
 
@@ -100,9 +119,9 @@ class OAuthProviderTest extends \PHPUnit_Framework_TestCase
         $roles = $result->getRoles();
         $this->assertCount(2, $roles);
         $this->assertInstanceOf('Symfony\Component\Security\Core\Role\Role', $roles[0]);
-        $this->assertEquals('ROLE_FOO', $roles[0]->getRole());
+        $this->assertSame('ROLE_FOO', $roles[0]->getRole());
         $this->assertInstanceOf('Symfony\Component\Security\Core\Role\Role', $roles[1]);
-        $this->assertEquals('ROLE_BAR', $roles[1]->getRole());
+        $this->assertSame('ROLE_BAR', $roles[1]->getRole());
     }
 
     public function testAuthenticateWithNullScope()
@@ -116,7 +135,8 @@ class OAuthProviderTest extends \PHPUnit_Framework_TestCase
         $this->serverService->expects($this->once())
             ->method('verifyAccessToken')
             ->with('x')
-            ->will($this->returnValue($accessToken));
+            ->will($this->returnValue($accessToken))
+        ;
 
         $result = $this->provider->authenticate($token);
 
@@ -138,7 +158,8 @@ class OAuthProviderTest extends \PHPUnit_Framework_TestCase
         $this->serverService->expects($this->once())
             ->method('verifyAccessToken')
             ->with('x')
-            ->will($this->returnValue($accessToken));
+            ->will($this->returnValue($accessToken))
+        ;
 
         $result = $this->provider->authenticate($token);
 
