@@ -1,5 +1,7 @@
 <?php
 
+declare(strict_types=1);
+
 /*
  * This file is part of the FOSOAuthServerBundle package.
  *
@@ -14,7 +16,6 @@ namespace FOS\OAuthServerBundle\DependencyInjection\Security\Factory;
 use Symfony\Bundle\SecurityBundle\DependencyInjection\Security\Factory\SecurityFactoryInterface;
 use Symfony\Component\Config\Definition\Builder\NodeDefinition;
 use Symfony\Component\DependencyInjection\ContainerBuilder;
-use Symfony\Component\DependencyInjection\DefinitionDecorator;
 use Symfony\Component\DependencyInjection\Reference;
 
 /**
@@ -29,16 +30,24 @@ class OAuthFactory implements SecurityFactoryInterface
      */
     public function create(ContainerBuilder $container, $id, $config, $userProvider, $defaultEntryPoint)
     {
+        // NOTE: done like this to avoid PHPStan complaining about a missing class for both Symfony v3 and Symfony v4
+        $definitionDecorator = 'Symfony\\Component\\DependencyInjection\\DefinitionDecorator';
+        $childDefinition = 'Symfony\\Component\\DependencyInjection\\ChildDefinition';
+        $definitionClass = $childDefinition;
+        if (class_exists($definitionDecorator)) {
+            $definitionClass = $definitionDecorator;
+        }
+
         $providerId = 'security.authentication.provider.fos_oauth_server.'.$id;
         $container
-            ->setDefinition($providerId, new DefinitionDecorator('fos_oauth_server.security.authentication.provider'))
+            ->setDefinition($providerId, new $definitionClass('fos_oauth_server.security.authentication.provider'))
             ->replaceArgument(0, new Reference($userProvider))
-            ;
+        ;
 
         $listenerId = 'security.authentication.listener.fos_oauth_server.'.$id;
-        $container->setDefinition($listenerId, new DefinitionDecorator('fos_oauth_server.security.authentication.listener'));
+        $container->setDefinition($listenerId, new $definitionClass('fos_oauth_server.security.authentication.listener'));
 
-        return array($providerId, $listenerId, 'fos_oauth_server.security.entry_point');
+        return [$providerId, $listenerId, 'fos_oauth_server.security.entry_point'];
     }
 
     /**
