@@ -1,5 +1,7 @@
 <?php
 
+declare(strict_types=1);
+
 /*
  * This file is part of the FOSOAuthServerBundle package.
  *
@@ -11,14 +13,14 @@
 
 namespace FOS\OAuthServerBundle;
 
-use FOS\OAuthServerBundle\DependencyInjection\Compiler\TokenStorageCompilerPass;
+use FOS\OAuthServerBundle\DependencyInjection\Compiler\GrantExtensionsCompilerPass;
 use FOS\OAuthServerBundle\DependencyInjection\Compiler\RequestStackCompilerPass;
+use FOS\OAuthServerBundle\DependencyInjection\Compiler\TokenStorageCompilerPass;
 use FOS\OAuthServerBundle\DependencyInjection\FOSOAuthServerExtension;
 use FOS\OAuthServerBundle\DependencyInjection\Security\Factory\OAuthFactory;
-use FOS\OAuthServerBundle\DependencyInjection\Compiler\GrantExtensionsCompilerPass;
+use Symfony\Bundle\SecurityBundle\DependencyInjection\SecurityExtension;
 use Symfony\Component\DependencyInjection\ContainerBuilder;
 use Symfony\Component\HttpKernel\Bundle\Bundle;
-use Symfony\Component\HttpKernel\Kernel;
 
 class FOSOAuthServerBundle extends Bundle
 {
@@ -31,10 +33,9 @@ class FOSOAuthServerBundle extends Bundle
     {
         parent::build($container);
 
-        if (version_compare(Kernel::VERSION, '2.1', '>=')) {
-            $extension = $container->getExtension('security');
-            $extension->addSecurityListenerFactory(new OAuthFactory());
-        }
+        /** @var SecurityExtension $extension */
+        $extension = $container->getExtension('security');
+        $extension->addSecurityListenerFactory(new OAuthFactory());
 
         $container->addCompilerPass(new GrantExtensionsCompilerPass());
         $container->addCompilerPass(new TokenStorageCompilerPass());
