@@ -1,5 +1,7 @@
 <?php
 
+declare(strict_types=1);
+
 /*
  * This file is part of the FOSOAuthServerBundle package.
  *
@@ -17,6 +19,9 @@ class TokenTest extends PropelTestCase
 {
     /**
      * @dataProvider getTestHasExpiredData
+     *
+     * @param mixed $expiresAt
+     * @param mixed $expect
      */
     public function testHasExpired($expiresAt, $expect)
     {
@@ -28,18 +33,18 @@ class TokenTest extends PropelTestCase
 
     public static function getTestHasExpiredData()
     {
-        return array(
-            array(time() + 60, false),
-            array(time() - 60, true),
-            array(null, false),
-        );
+        return [
+            [time() + 60, false],
+            [time() - 60, true],
+            [null, false],
+        ];
     }
 
     public function testExpiresIn()
     {
         $token = new Token();
 
-        $this->assertEquals(PHP_INT_MAX, $token->getExpiresIn());
+        $this->assertSame(PHP_INT_MAX, $token->getExpiresIn());
     }
 
     public function testExpiresInWithExpiresAt()
@@ -47,7 +52,7 @@ class TokenTest extends PropelTestCase
         $token = new Token();
         $token->setExpiresAt(time() + 60);
 
-        $this->assertEquals(60, $token->getExpiresIn());
+        $this->assertSame(60, $token->getExpiresIn());
     }
 }
 
