@@ -1,5 +1,7 @@
 <?php
 
+declare(strict_types=1);
+
 /*
  * This file is part of the FOSOAuthServerBundle package.
  *
@@ -26,26 +28,35 @@ class OAuthListenerTest extends TestCase
 
     public function setUp()
     {
-        $this->serverService = $this
-            ->getMockBuilder('OAuth2\OAuth2')
+        $this->serverService = $this->getMockBuilder('OAuth2\OAuth2')
             ->disableOriginalConstructor()
-            ->getMock();
+            ->getMock()
+        ;
 
         $this->authManager = $this
-            ->getMock('Symfony\Component\Security\Core\Authentication\AuthenticationManagerInterface');
+            ->getMockBuilder('Symfony\Component\Security\Core\Authentication\AuthenticationManagerInterface')
+            ->disableOriginalConstructor()
+            ->getMock()
+        ;
 
         if (interface_exists('Symfony\Component\Security\Core\Authentication\Token\Storage\TokenStorageInterface')) {
             $this->securityContext = $this
-                ->getMock('Symfony\Component\Security\Core\Authentication\Token\Storage\TokenStorageInterface');
+                ->getMockBuilder('Symfony\Component\Security\Core\Authentication\Token\Storage\TokenStorageInterface')
+                ->disableOriginalConstructor()
+                ->getMock()
+            ;
         } else {
-            $this->securityContext = $this
-                ->getMock('Symfony\Component\Security\Core\SecurityContextInterface');
+            $this->securityContext = $this->getMockBuilder('Symfony\Component\Security\Core\SecurityContextInterface')
+                ->disableOriginalConstructor()
+                ->getMock()
+            ;
         }
 
         $this->event = $this
             ->getMockBuilder('Symfony\Component\HttpKernel\Event\GetResponseEvent')
             ->disableOriginalConstructor()
-            ->getMock();
+            ->getMock()
+        ;
     }
 
     public function testHandle()
@@ -55,22 +66,25 @@ class OAuthListenerTest extends TestCase
         $this->serverService
             ->expects($this->once())
             ->method('getBearerToken')
-            ->will($this->returnValue('a-token'));
+            ->will($this->returnValue('a-token'))
+        ;
 
         $this->authManager
             ->expects($this->once())
             ->method('authenticate')
-            ->will($this->returnArgument(0));
+            ->will($this->returnArgument(0))
+        ;
 
         $this->securityContext
             ->expects($this->once())
             ->method('setToken')
-            ->will($this->returnArgument(0));
+            ->will($this->returnArgument(0))
+        ;
 
         $token = $listener->handle($this->event);
 
         $this->assertInstanceOf('FOS\OAuthServerBundle\Security\Authentication\Token\OAuthToken', $token);
-        $this->assertEquals('a-token', $token->getToken());
+        $this->assertSame('a-token', $token->getToken());
     }
 
     public function testHandleResponse()
@@ -80,26 +94,33 @@ class OAuthListenerTest extends TestCase
         $this->serverService
             ->expects($this->once())
             ->method('getBearerToken')
-            ->will($this->returnValue('a-token'));
+            ->will($this->returnValue('a-token'))
+        ;
 
-        $response = $this->getMock('Symfony\Component\HttpFoundation\Response');
+        $response = $this->getMockBuilder('Symfony\Component\HttpFoundation\Response')
+            ->disableOriginalConstructor()
+            ->getMock()
+        ;
 
         $this->authManager
             ->expects($this->once())
             ->method('authenticate')
-            ->will($this->returnValue($response));
+            ->will($this->returnValue($response))
+        ;
 
         $this->securityContext
             ->expects($this->never())
-            ->method('setToken');
+            ->method('setToken')
+        ;
 
         $this->event
             ->expects($this->once())
             ->method('setResponse')
-            ->will($this->returnArgument(0));
+            ->will($this->returnArgument(0))
+        ;
 
         $ret = $listener->handle($this->event);
 
-        $this->assertEquals($response, $ret);
+        $this->assertSame($response, $ret);
     }
 }
